feat: move animation logic to base
